adc2, adc3 testing, new cpis testing

+++ Vcu_Config.h
@@ -358,7 +358,9 @@ XMC_RTC_TIME_t current_time,timeval;
 GPS_data_t GPS_data_val;
 time_t epoch;
 // Adding for ambient_temp
-uint8_t ambient_temp;
+float ambient_temp;
 
 //flag added to check if vehicle is in drive mode or not (i.e ign 1) used for packet publishing using 4G
 bool is_drive_mode;
+
+//float U18_read_channel(uint8_t ch_no);
